pii-guard/requirements.txt
# pii-guard/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
PyPDF2>=3.0.0
pyyaml>=6.0.0